requirements.txt
aniso8601==8.0.0
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
dnspython==2.0.0
Flask==1.1.2
Flask-Cors==3.0.9
Flask-RESTful==0.3.8
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
jsonify==0.5
MarkupSafe==1.1.1
nltk==3.5
pymongo==3.11.1
python-dotenv==0.15.0
pytz==2020.4
regex==2020.11.13
requests==2.25.0
schedule==0.6.0
six==1.15.0
textblob==0.15.3
tqdm==4.52.0
urllib3==1.26.2
Werkzeug==1.0.1
